requirements.txt
# Core training and model dependencies
einops==0.6.1           # updated for modern compatibility
matplotlib==3.7.2       # works on Py3.10+
numpy==1.26.4           # latest stable, fixes the C-extension issues
opencv_python==4.8.1.78 # modern, ARM64 compatible
Pillow==9.5.0           # modern Pillow compatible with M2
scipy==1.11.1           # compatible with numpy 1.26
timm==0.9.2             # updated to work with torch >= 2.1
torch>=2.1.0            # latest ARM64 version for M2
torchvision>=0.15.2     # matching torch version
tqdm==4.66.1             # latest stable
wandb==0.15.8           # updated

# For config parsing and logging
PyYAML==6.0
rich==13.5.2

# For evaluation, inference loss, and metrics
scikit-image>=0.20.0    # PSNR, SSIM, structural metrics
imageio>=2.31.0         # alternative image I/O
psutil>=5.9.5           # system / memory monitoring

# Optional visualization and analysis
seaborn>=0.12.2
setuptools>=68.0.0
packaging>=23.0
